feat(technologies): add begin and end date for each technology
- add begin and end date for technologies
-- add validator for properties mentioned above

- home page rework
-- add greeting text
-- add notifications panel

- new EF migration
-- HourLength type change from int to decimal
-- IsDeleted EF model refactoring

- minor style tweaks
- minor site.js refactoring

+++ NPO_Workflow/ViewModels/Operations/OperationViewModel.cs
@@ -21,6 +21,6 @@ namespace NPO_Workflow.ViewModels.Operations
 
         [Required(ErrorMessage = "Длительность операции обязательна")]
         [Range(0, int.MaxValue, ErrorMessage = "Длительность не может быть отрицательной")]
-        public int? HourLength { get; set; }
+        public decimal? HourLength { get; set; }
     }
 }
